power: Enable sched_mask_force for user space library.
Some userspace applications (for games) uses libraries,
which query the cpu topology and attach to the highest
clocked cpus. In case the soc provides same clock rates
on both big.LITTLE clusters it's not guranteed that the
app is affined to the cluster with the most capacities.

In high performance profiles, this sets the
neccessary kernel options, while in balances/low-power
profiles this is deactivated.

Change-Id: I7d192e93231629c1cfe08c34552f45eaac85d9bd